Correct type id_user

+++ src/main/java/com/ihc/makaraBack/controller/UsuarioController.java
@@ -19,7 +19,7 @@ public class UsuarioController {
     }
 
     @GetMapping(value = "/usuario")
-    public Usuario getById(Integer id){
+    public Usuario getById(Long id){
         return usuarioService.getById(id);
     }
